Move `disk_params` and `disk_config_params` to their own unit
Export these structs so they can be used from within tests
to configure an io_queue and avoid duplicating code already
written in these types.
